[luci/service] Migrate shape inference for Conv2D, DepthwiseConv2D
This commit migrates shape inference of Conv2D and DepthwiseConv2D to sinf::Algorithm.

ONE-DCO-1.0-Signed-off-by: Chansu Park <1265pcs@gmail.com>
